% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datapoints.R
\name{create_datapoints}
\alias{create_datapoints}
\title{Create datapoint-files}
\usage{
create_datapoints(
  dataset,
  entity_id,
  variable_id,
  datacolumns,
  output_folder = getwd()
)
}
\arguments{
\item{dataset}{A data frame. The data set to be converted. Data has to be in wide-format.}

\item{entity_id}{A character variable. The ID connected to each entity. Should only include lower case letters and underscores.}

\item{variable_id}{A character variable. The ID connected to each variable. Should only include lower case letters and underscores.}

\item{datacolumns}{A range of numeric variables. The columns including data. Should be referenced in the format "1-3" or "1,2,3"}

\item{output_folder}{(Optional) Where the file should be saved. By default set to the working directory.}
}
\value{
Saves a folder in the given output-folder named "ddf--datapoints". This folder includes one CSV-file for each unique variable in the data set.
}
\description{
This function creates one datapoint-file for each variable in the data set in the DDF-format
}
\examples{
\dontrun{
create_datapoints(dataset = dataset, variable_id = "var_id", entity_id = "reg_id",
datacolumns = "7-21", output_folder = get_wd())
}
}
