% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tags.R
\name{create_tags}
\alias{create_tags}
\title{Create Tag-file}
\usage{
create_tags(dataset, tag_id, tag_name, parent = NULL, output_folder = getwd())
}
\arguments{
\item{dataset}{A data frame. The data set to be converted. Data has to be in wide-format.}

\item{tag_id}{A character variable. The ID connected to each variable. Should only include lower case letters and underscores.}

\item{tag_name}{A character variable. The Name of each variable.}

\item{parent}{(Optional) A character variable. The definition of each variable. Should not include commas or semi-colons.}

\item{output_folder}{(Optional) Where the file should be saved. By default set to the working directory.}
}
\value{
Saves a CSV-file in the given output-folder named "ddf--entities--tag". This file includes one row for each named tag ID with name, and optionally parent tag.
}
\description{
This function creates a tag-file in the DDF-format
}
\examples{
\dontrun{
create_tags(dataset = dataset, tag_id = "tag_id", tag_name = "Name",
 output_folder = "path/to/folder")
}
}
