% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_concepts.R
\name{create_concepts}
\alias{create_concepts}
\title{Create Concept-file}
\usage{
create_concepts(
  dataset,
  variable_id,
  variable_name,
  entity_id,
  description = NULL,
  source = NULL,
  name_short = NULL,
  name_catalog = NULL,
  tag_id = NULL,
  output_folder = getwd(),
  return_output = FALSE
)
}
\arguments{
\item{dataset}{A data frame. The data set to be converted. Data has to be in long-format.}

\item{variable_id}{A character variable. The ID connected to each variable. Should only include lower case letters and underscores.}

\item{variable_name}{A character variable. The Name of each variable.}

\item{entity_id}{A character variable. The ID connected to each entity. Should only include lower case letters and underscores.}

\item{description}{(Optional) A character variable. The definition of each variable. Should not include commas or semi-colons.}

\item{source}{(Optional) A character variable. The source of each variable. Preferably the URL of the source.}

\item{name_short}{(Optional) A character variable. A shortened version of the variable name.}

\item{name_catalog}{(Optional) A character variable. The variable name to be shown in the catalog / tree menu.}

\item{tag_id}{(Optional) A character variable. The ID specifying which folder the variable belongs in. Should only include lower case letters and underscores.}

\item{output_folder}{(Optional) Where the file should be saved. By default set to the working directory.}

\item{return_output}{Returns the output. Default = FALSE}
}
\value{
Saves a CSV-file in the given output-folder named "ddf--concepts". This file includes one row for each named variable which always includes geo, year, name, domain, entity_id, all data-variables and the given, relevant variables.
}
\description{
This function creates a concepts-file in the DDF-format
}
\examples{
\dontrun{
create_concepts(dataset = dataset, variable_id = "var_id", variable_name = "Variable",
entity_id = "reg_id", source = "source_url")
}
}
