% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_DDF.R
\name{create_ddf}
\alias{create_ddf}
\title{Create DDF-filesystem}
\usage{
create_ddf(
  dataset,
  variable_id,
  variable_name,
  entity_id,
  entity_name,
  datacolumns,
  tag_id,
  tag_name,
  name_short = NULL,
  name_catalog = NULL,
  description = NULL,
  source = NULL,
  parent = NULL,
  output_folder = getwd()
)
}
\arguments{
\item{dataset}{A data frame. The data set to be converted. Data has to be in wide-format.}

\item{variable_id}{A character variable. The ID connected to each variable. Should only include lower case letters and underscores.}

\item{variable_name}{A character variable. The Name of each variable.}

\item{entity_id}{A character variable. The ID connected to each entity. Should only include lower case letters and underscores.}

\item{entity_name}{A character variable. The Name of each entity.}

\item{datacolumns}{A range of numeric variables. The columns including data. Should be referenced in the format "1-3" or "1,2,3"}

\item{tag_id}{A character variable. The ID connected to each variable. Should only include lower case letters and underscores.}

\item{tag_name}{A character variable. The Name of each variable.}

\item{name_short}{(Optional) A character variable. A shortened version of the variable name.}

\item{name_catalog}{(Optional) A character variable. The variable name to be shown in the catalog / tree menu.}

\item{description}{(Optional) A character variable. The definition of each variable. Should not include commas or semi-colons.}

\item{source}{(Optional) A character variable. The source of each variable. Preferably the URL of the source.}

\item{parent}{(Optional) A character variable. The definition of each variable. Should not include commas or semi-colons.}

\item{output_folder}{(Optional) Where the file should be saved. By default set to the working directory.}
}
\value{
Saves a folder in the given output-folder named "ddf_filesystem_current_date". This folder includes a concepts-files, entities-file and a folder containing all datapoints.
}
\description{
This function creates a filesystem in the DDF-format
}
\examples{
\dontrun{
create_ddf(dataset = dataset, variable_id = "var_id", variable_name = "Variable",
entity_id = "reg_id", entity_name = "Region", datacolumns = "7-21", description = "description",
output_folder = get_wd())
}
}
